pemograman-kripto/05-pertemuan-5/requirement.txt
logging==0.4.9.6
psycopg2-binary==2.9.11
python-dotenv==1.2.2
requests==2.32.5
